+++ packages/api-security-admin-users/src/plugins/APIKeyAuthorizationPlugin.ts
@@ -1,4 +1,3 @@
-import { SecurityIdentity, SecurityPermission } from "@webiny/api-security/types";
 import { AdminUsersContext } from "~/types";
 import { AuthorizationPlugin } from "@webiny/api-security/plugins/AuthorizationPlugin";
 
@@ -6,12 +5,7 @@ export interface Config {
     identityType?: string;
 }
 
-const createCacheKey = ({ identity }: { identity: SecurityIdentity }): string => {
-    return `I#${identity.id}`;
-};
-
 export class APIKeyAuthorizationPlugin extends AuthorizationPlugin<AdminUsersContext> {
-    private _permissionCache = new Map<string, SecurityPermission[] | null>();
     private _config: Config;
 
     constructor(config?: Config) {
@@ -27,17 +21,9 @@ export class APIKeyAuthorizationPlugin extends AuthorizationPlugin<AdminUsersCon
         if (!identity || identity.type !== identityType) {
             return;
         }
-        const cacheKey = createCacheKey({ identity });
 
-        if (this._permissionCache.has(cacheKey)) {
-            return this._permissionCache.get(cacheKey);
-        }
         // We can expect `permissions` to exist on the identity, because api-key authentication
         // plugin sets them on the identity instance to avoid loading them from DB here.
-        const permissions = Array.isArray(identity.permissions) ? identity.permissions : [];
-
-        this._permissionCache.set(cacheKey, permissions);
-
-        return permissions;
+        return Array.isArray(identity.permissions) ? identity.permissions : [];
     }
 }