+++ packages/app-admin/src/components/FileManager/FileManagerView.tsx
@@ -392,6 +392,7 @@ function FileManagerView(props: FileManagerViewProps) {
                                 onChange={e => searchOnChange(e.target.value)}
                                 placeholder={t`Search by filename or tags`}
                                 disabled={!fmFilePermission}
+                                data-testid={"file-manager.search-input"}
                             />
                         </InputSearch>
                     }