+++ packages/app-admin/src/ui/elements/ViewElement.tsx
@@ -1,0 +1,1 @@
+export * from "@webiny/ui-composer/elements/ViewElement";