+++ packages/app-form-builder/src/admin/plugins/formDetails/formRevisions/Revision.tsx
@@ -69,7 +69,10 @@ const Revision = (props: RevisionProps) => {
         deleteRevision,
         editRevision,
         unpublishRevision
-    } = useRevision({ revision: rev, form });
+    } = useRevision({
+        revision: rev,
+        form
+    });
     const { canPublish, canUnpublish, canDelete, canEdit } = usePermission();
 
     const showMenu = canEdit(form) || canDelete(form) || canPublish() || canUnpublish();